% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeAtlantisInputs.R
\name{writeAtlantisInputs}
\alias{writeAtlantisInputs}
\title{Construct a matrix and an annotation table for a set of matrices for use with ATLANTIS}
\usage{
writeAtlantisInputs(output.dir, pred.sources, target.sources,
  loci.by.symbol = NULL, analysisID = "myAnalysis")
}
\arguments{
\item{output.dir}{The directory to write the two matrices as Rdata files}

\item{pred.sources}{The definitions of the predictive features to use (See below)}

\item{loci.by.symbol}{A character vector where the elements are named by gene symbols}

\item{target.source}{The definitions of the target matrix to use (See below)}
}
\description{
Construct a matrix and an annotation table for a set of matrices for use with ATLANTIS
}
\section{Specifying data sources}{

Both pred.sources and target.source are defined as a list of definitions where each definition has 
the following fields: dat, prefix, vartype, must.have.cellline

The 'dat' field is the numeric matrix of values.  'prefix' is a short code to prepend to each feature name from that matrix.  (For example "CN" for copy number to differentiate the feature from the gene expression for the same gene.)
'vartype' must be one of 'MUT', 'Exp', 'CN' and controls how the values will be plotted.   Lastly, must.have.cellline controls whether NAs should be used to fill in features that are missing a 
cell line from this dataset, or whether the cell line should be dropped from the others if its missing.

Upon completion, returns a list with the following fields which point to the generated files: target.file, pred.file, anno.file
}

